@stevenh 2022-07-01
Looks like the proposal https://github.com/golang/go/issues/53003 has been accepted recently so we can look to use those when they are implemented.